requirements.txt
torch
transformers
sentencepiece
deepgram-sdk
python-dotenv
demucs
google-generativeai
elevenlabs
